refactor(cli): centralizar clasificacion de ParserError en helper compartido
